requirements.txt
requests
python-dotenv
qrcode[pil]
schedule
PyYAML